+++ Engine/PluginSearcher.cs
@@ -375,10 +375,7 @@ namespace OpenTap
             return PluginTypes;
         }
 
-        internal TypeData PluginMarkerType = new TypeData
-        {
-            Name = typeof(ITapPlugin).FullName
-        };
+        internal readonly TypeData PluginMarkerType = new TypeData(typeof(ITapPlugin).FullName);
 
         private void PluginsInAssemblyRecursive(AssemblyData asm)
         {
@@ -504,11 +501,12 @@ namespace OpenTap
                     return null;
             }
 
-            TypeData plugin = new TypeData();
+            string typeName;
+            
             TypeDefinitionHandle declaringTypeHandle = typeDef.GetDeclaringType();
             if (declaringTypeHandle.IsNil)
             {
-                plugin.Name = string.Format("{0}.{1}", CurrentReader.GetString(typeDef.Namespace), CurrentReader.GetString(typeDef.Name));
+                typeName = string.Format("{0}.{1}", CurrentReader.GetString(typeDef.Namespace), CurrentReader.GetString(typeDef.Name));
             }
             else
             {
@@ -516,11 +514,10 @@ namespace OpenTap
                 TypeData declaringType = PluginFromTypeDefRecursive(declaringTypeHandle);
                 if (declaringType == null)
                     return null;
-                plugin.Name = string.Format("{0}+{1}", declaringType.Name, CurrentReader.GetString(typeDef.Name));
+                typeName = string.Format("{0}+{1}", declaringType.Name, CurrentReader.GetString(typeDef.Name));
             }
-            if (AllTypes.ContainsKey(plugin.Name))
+            if (AllTypes.TryGetValue(typeName, out var existingPlugin))
             {
-                var existingPlugin = AllTypes[plugin.Name];
                 if (existingPlugin.Assembly.Name == CurrentAsm.Name)
                 {
                     // we assume this is the same plugin, just in another copy of the dll
@@ -536,6 +533,7 @@ namespace OpenTap
                 }
                 return existingPlugin;
             }
+            TypeData plugin = new TypeData(typeName);
             if (plugin.Name == PluginMarkerType.Name)
             {
                 PluginMarkerType.Assembly = CurrentAsm;
@@ -865,12 +863,12 @@ namespace OpenTap
 
             public TypeData GetPrimitiveType(PrimitiveTypeCode typeCode)
             {
-                return new TypeData { Name = "System." + typeCode.ToString() };
+                return new TypeData("System." + typeCode);
             }
 
             public TypeData GetSZArrayType(TypeData elementType)
             {
-                return elementType != null ? new TypeData { Name = elementType.Name } : null;
+                return elementType != null ? new TypeData(elementType.Name) : null;
             }
 
             public TypeData GetTypeFromDefinition(MetadataReader reader, TypeDefinitionHandle handle, byte rawTypeKind)
@@ -889,181 +887,6 @@ namespace OpenTap
             }
         }
         #endregion
-    }
-
-    /// <summary>
-    /// Representation of a C#/dotnet type including its inheritance hierarchy. Part of the object model used in the PluginManager
-    /// </summary>
-    [DebuggerDisplay("{Name}")]
-    public partial class TypeData
-    {
-        /// <summary>
-        /// Gets the fully qualified name of the type, including its namespace but not its assembly.
-        /// </summary>
-        public string Name { get; internal set; }
-
-        /// <summary>
-        /// Gets the TypeAttributes for this type. This can be used to check if the type is abstract, nested, an interface, etc.
-        /// </summary>
-        public TypeAttributes TypeAttributes { get; internal set; }
-        
-        /// <summary>
-        /// Gets the Assembly that defines this type.
-        /// </summary>
-        public AssemblyData Assembly { get; internal set; }
-
-        // Used to mark when no display attribute is present.
-        static readonly DisplayAttribute noDisplayAttribute = new DisplayAttribute("<<Null>>");
-        
-        DisplayAttribute display;
-        /// <summary>
-        /// Gets.the DisplayAttribute for this type. Null if the type does not have a DisplayAttribute
-        /// </summary>
-        public DisplayAttribute Display
-        {
-            get
-            {
-                if (display is null && attributes != null)
-                {
-                    display = noDisplayAttribute;
-                    foreach (var attr in attributes)
-                    {
-                        if (attr is DisplayAttribute displayAttr)
-                        {
-                            display = displayAttr;
-                            break;
-                        }
-                    }
-                    
-                }
-
-                if (ReferenceEquals(display, noDisplayAttribute)) 
-                    return null;
-                return display;
-            }
-            internal set => display = value;
-        }
-
-
-        ICollection<TypeData> baseTypes;
-        
-        /// <summary> Gets a list of base types (including interfaces) </summary>
-        internal ICollection<TypeData> BaseTypes => baseTypes;
-
-        internal void FinalizeCreation()
-        {
-            baseTypes = baseTypes?.ToArray();
-            pluginTypes = pluginTypes?.ToArray();
-        }
-        internal void AddBaseType(TypeData typename)
-        {
-            if (baseTypes == null)
-                baseTypes = new HashSet<TypeData>();
-            baseTypes.Add(typename);
-        }
-
-        ICollection<TypeData> pluginTypes;
-        /// <summary>
-        /// Gets a list of plugin types (i.e. types that directly implement ITapPlugin) that this type inherits from/implements
-        /// </summary>
-        public IEnumerable<TypeData> PluginTypes => pluginTypes;
-
-        internal void AddPluginType(TypeData typename)
-        {
-            if (typename == null)
-                return;
-            if (pluginTypes == null)
-                pluginTypes = new HashSet<TypeData>();
-            pluginTypes.Add(typename);
-        }
-        internal void AddPluginTypes(IEnumerable<TypeData> types)
-        {
-            if (types == null)
-                return;
-            if (pluginTypes == null)
-                pluginTypes = new HashSet<TypeData>();
-            foreach (var t in types)
-                pluginTypes.Add(t);
-        }
-
-        ICollection<TypeData> derivedTypes;
-
-        /// <summary>
-        /// Gets a list of types that has this type as a base type (including interfaces)
-        /// </summary>
-        public IEnumerable<TypeData> DerivedTypes => derivedTypes ?? Array.Empty<TypeData>();
-
-        /// <summary>
-        /// False if the type has a System.ComponentModel.BrowsableAttribute with Browsable = false.
-        /// </summary>
-        public bool IsBrowsable { get; internal set; }
-
-        internal void AddDerivedType(TypeData typename)
-        {
-            if (derivedTypes == null)
-                derivedTypes = new HashSet<TypeData>();
-            else if (derivedTypes.Contains(typename))
-                return;
-            derivedTypes.Add(typename);
-            if (BaseTypes != null)
-            {
-                foreach (TypeData b in BaseTypes)
-                    b.AddDerivedType(typename);
-            }
-        }
-
-        internal TypeData()
-        {
-            IsBrowsable = true;
-        }
-        
-        private bool failedLoad;
-
-        /// <summary>
-        /// Returns the System.Type corresponding to this. 
-        /// If the assembly in which this type is defined has not yet been loaded, this call will load it.
-        /// </summary>
-        public Type Load()
-        {
-            if (failedLoad) return null;
-            if (type != null) return type;
-
-            try
-            {
-                var asm = Assembly.Load();
-                if (asm == null)
-                {
-                    failedLoad = true;
-                    return null;
-                }
-
-                type = asm.GetType(this.Name, true);
-                dict.GetValue(type, t => this);
-            }
-            catch (Exception ex)
-            {
-                failedLoad = true;
-                log.Error("Unable to load type '{0}' from '{1}'. Reason: '{2}'.", Name, Assembly.Location,
-                    ex.Message);
-                log.Debug(ex);
-            }
-
-            return type;
-        }
-
-        /// <summary> The loaded state of the type. </summary>
-        internal LoadStatus Status => type != null ? LoadStatus.Loaded : (failedLoad ? LoadStatus.FailedToLoad : LoadStatus.NotLoaded);
-
-        static TraceSource log = Log.CreateSource("PluginManager");
-
-        /// <summary>
-        /// Returns the DisplayAttribute.Name if the type has a DisplayAttribute, otherwise the FullName without namespace
-        /// </summary>
-        /// <returns></returns>
-        public string GetBestName()
-        {
-            return Display != null ? Display.Name : Name.Split('.', '+').Last();
-        }
     }
 
     /// <summary>
