+++ Engine/Reflection/DotNetTypeData.cs
@@ -13,327 +13,6 @@ using System.Runtime.CompilerServices;
 
 namespace OpenTap
 {
-    /// <summary> Represents a .NET type. </summary>
-    public partial class TypeData : ITypeData
-    {
-
-        /// <summary> Creates a string value of this.</summary>
-        public override string ToString() => Name;
-
-        static ConditionalWeakTable<Type, TypeData> dict = new ConditionalWeakTable<Type, TypeData>();
-
-        Type type;
-
-        /// <summary>
-        /// Gets the System.Type that this represents. Same as calling <see cref="Load()"/>.
-        /// Accessing this property causes the underlying Assembly to be loaded if it is not already.
-        /// </summary>
-        public Type Type => Load();
-
-        // add assembly is not thread safe.
-        static object loadTypeDictLock = new object();
-        
-        /// <summary> Creates a new TypeData object to represent a dotnet type. </summary>
-        public static TypeData FromType(Type type)
-        {
-            checkCacheValidity();
-            if (dict.TryGetValue(type, out var i))
-                return i;
-            TypeData td = null;
-            lock (loadTypeDictLock)
-            {
-                var searcher = PluginManager.GetSearcher();
-
-                searcher?.AllTypes.TryGetValue(type.FullName, out td);
-                if (td == null && searcher != null)
-                {
-                    // This can occur for some types inside mscorlib such as System.Net.IPAddress.
-                    try
-                    {
-                        if (type.Assembly != null && type.Assembly.IsDynamic == false &&
-                            type.Assembly.Location != null)
-                        {
-                            searcher.AddAssembly(type.Assembly.Location, type.Assembly);
-                            if (searcher.AllTypes.TryGetValue(type.FullName, out td))
-                                return td;
-                        }
-                    }
-                    catch
-                    {
-                    }
-
-                    td = new TypeData(type);
-                }
-                else
-                {
-                    // This can occur when using shared projects because the same type is defined in multiple different assemblies with the same fully qualified
-                    // name (namespace+typename). In this case, it is possible for the PluginSearcher to resolve an instance of this typedata from a different
-                    // type than the input type to this method, which will cause all sorts of reflection errors. We detect this edge case here, and if there
-                    // is a type mismatch, we instantiate a new typedata from the correct type.
-                    if (td == null || td.Type != type)
-                    {
-                        td = new TypeData(type);
-                    }
-                }
-            }
-
-            return dict.GetValue(type, x => td);
-        }
-        
-        TypeData(Type type)
-        {
-            this.type = type;
-            this.Name = type.FullName;
-            postload();
-        }
-
-        bool postLoaded = false;
-        object loadLock = new object();
-        void postload()
-        {
-            if (postLoaded) return;
-            lock (loadLock)
-            {
-                if (postLoaded) return;
-                Load();
-                if (type != typeof(string))
-                {
-                    var elementType = type.GetEnumerableElementType();
-                    if (elementType != null)
-                        this.elementType = FromType(elementType);
-                }
-                typecode = Type.GetTypeCode(type);
-                hasFlags = this.HasAttribute<FlagsAttribute>();
-                isValueType = type.IsValueType;
-                postLoaded = true;
-            }
-        }
-
-        bool isValueType;
-        /// <summary> gets if the type is a value-type. (see Type.IsValueType)</summary>
-        public bool IsValueType
-        {
-            get
-            {
-                postload();
-                return isValueType;
-            }
-        }
-
-        internal bool IsNumeric
-        {
-            get
-            {
-                postload();
-                if (type.IsEnum)
-                    return false;
-                switch (typecode)
-                {
-                    case TypeCode.Byte:
-                    case TypeCode.SByte:
-                    case TypeCode.UInt16:
-                    case TypeCode.UInt32:
-                    case TypeCode.UInt64:
-                    case TypeCode.Int16:
-                    case TypeCode.Int32:
-                    case TypeCode.Int64:
-                    case TypeCode.Decimal:
-                    case TypeCode.Double:
-                    case TypeCode.Single:
-                        return true;
-                    default:
-                        return false;
-                }
-            }
-        }
-
-        internal bool IsString
-        {
-            get
-            {
-                postload();
-                return typecode == TypeCode.String;
-            }
-        }
-        TypeCode typecode = TypeCode.Object;
-
-        object[] attributes = null;
-        /// <summary> 
-        /// The attributes of this type. 
-        /// Accessing this property causes the underlying Assembly to be loaded if it is not already.
-        /// </summary>
-        public IEnumerable<object> Attributes => attributes ?? (attributes = Load()?.GetAllCustomAttributes(false)) ?? Array.Empty<object>();
-
-        ITypeData baseTypeCache;
-        /// <summary> The base type of this type. Will return null if there is no base type. If there is no direct base, but instead an interface, that will be returned.</summary>
-        public ITypeData BaseType
-        {
-            get
-            {
-                if (baseTypeCache != null) return ReferenceEquals(baseTypeCache, NullTypeData.Instance) ? null : baseTypeCache;
-                baseTypeCache = BaseTypes?.ElementAtOrDefault(0);
-                if (baseTypeCache != null) return baseTypeCache;
-                var result2 = Load()?.BaseType;
-                if (result2 != null)
-                {
-                    baseTypeCache = FromType(result2);
-                    return baseTypeCache;
-                }
-                baseTypeCache = NullTypeData.Instance;
-                return baseTypeCache;
-            }
-        }
-
-        TypeData elementType;
-
-        /// <summary> If this is a collection type, then this is the element type. Otherwise null. </summary>
-        internal TypeData ElementType
-        {
-            get
-            {
-                postload();
-                return elementType;
-            }
-        }
-
-        internal bool createInstanceSet => canCreateInstance.HasValue;
-        
-        bool? canCreateInstance;
-        /// <summary> 
-        /// returns true if an instance possibly can be created. 
-        /// Accessing this property causes the underlying Assembly to be loaded if it is not already.
-        /// </summary>
-        public bool CanCreateInstance {
-            get
-            {
-                if (failedLoad) return false;
-                if (canCreateInstance.HasValue) return canCreateInstance.Value;
-                if (Load() is Type t)
-                {
-                    type = t;
-                    canCreateInstance = type.IsAbstract == false && type.IsInterface == false &&
-                                        type.ContainsGenericParameters == false &&
-                                        type.GetConstructor(Array.Empty<Type>()) != null;
-                    return canCreateInstance.Value;
-                }
-                return false; // failed to load
-            }
-            internal set => canCreateInstance = value;
-        }
-
-        string assemblyQualifiedName;
-        internal string AssemblyQualifiedName
-        {
-            get
-            {
-                if (failedLoad) return "";
-                return assemblyQualifiedName ?? (assemblyQualifiedName = Load().AssemblyQualifiedName);
-            }
-        }
-
-        /// <summary>
-        /// Creates a new object instance of this type.
-        /// Accessing this property causes the underlying Assembly to be loaded if it is not already.
-        /// </summary>
-        public object CreateInstance(object[] arguments)
-        {
-            if (!(Load() is Type t))
-                throw new InvalidOperationException(
-                    $"Failed to instantiate object of type '{this.Name}'. The assembly failed to load.");
-            return Activator.CreateInstance(t, arguments);
-        }
-
-        /// <summary>
-        /// Gets a member by name.
-        /// Causes the underlying Assembly to be loaded if it is not already.
-        /// </summary>
-        public IMemberData GetMember(string name)
-        {
-            var members = (IMemberData[]) GetMembers();
-            foreach(var member in members)
-            {
-                if(member.Name == name)
-                    return member;
-            }
-            return null;
-        }
-        
-        IMemberData[] members;
-
-        /// <summary>
-        /// Gets all the members of this type. 
-        /// Causes the underlying Assembly to be loaded if it is not already.
-        /// </summary>
-        public IEnumerable<IMemberData> GetMembers()
-        {
-            if (members != null) return members;
-
-            if (Load() is Type t)
-            {
-                var props = t.GetPropertiesTap();
-                List<IMemberData> m = new List<IMemberData>(props.Length);
-                foreach (var mem in props)
-                {
-                    try
-                    {
-                        if (mem.GetMethod != null && mem.GetMethod.GetParameters().Length > 0)
-                            continue;
-
-                        if (mem.SetMethod != null && mem.SetMethod.GetParameters().Length != 1)
-                            continue;
-                    }
-                    catch
-                    {
-                        continue;
-                    }
-
-                    m.Add(MemberData.Create(mem));
-                }
-
-                foreach (var mem in t.GetMethodsTap())
-                {
-                    if (mem.GetAttribute<BrowsableAttribute>()?.Browsable ?? false)
-                    {
-                        var member = MemberData.Create(mem);
-                        m.Add(member);
-                    }
-                }
-                members = m.ToArray();
-            }
-            else
-            {
-                // The members list cannot be populated because the type could not be loaded.
-                members = Array.Empty<IMemberData>();
-            }
-            return members;
-        }
-
-        bool hasFlags;
-        internal bool HasFlags()
-        {
-            postload();
-            return hasFlags;
-        }
-
-        /// <summary> Compares two TypeDatas by comparing their inner Type instances. </summary>
-        /// <param name="obj"> Should be a TypeData</param>
-        /// <returns>true if the two Type properties are equals.</returns>
-        public override bool Equals(object obj)
-        {
-            if (obj is TypeData td && td.type != null && type != null )
-                return td.type == type;
-            return ReferenceEquals(this, obj);
-        }
-
-        /// <summary> Calculates the hash code based on the .NET Type instance. </summary>
-        /// <returns></returns>
-        public override int GetHashCode()
-        {
-            var asm = Assembly?.GetHashCode() ?? 0;
-            return (asm + 586093897) * 759429795 + 
-                   (Name.GetHashCode() + 836431542) * 678129773;
-        }
-    }
 
     /// <summary>
     /// Represents the members of C#/dotnet types.
